Implement async methods.
For sake of performance consideration we need to implement this asynchronous
methods. For heavy loaded web site we don't want to hold a thread on the
potentially long running operation as a remote method invocation.

This commit contains an example on how to use these async methods in
context of ASP.NET async page execution.