Implement Session storage provider.
Saving to ASP.NET session is cool. But for some scenarious it is useful
to keep track of user's session by means of cookies. The new connect
library provides us with a mechanism to login/auth and to save session
into a cookie with name 'fbs_' + appId.

With this storage provider implementation developers can pick with
approach to re-use, or implement their own provider.